+++ js/data.js
@@ -1,0 +1,3 @@
+import {createImage, PHOTOS_COUNT} from './data-images.js';
+
+Array.from( {length: PHOTOS_COUNT}, createImage);
